Update class name due to internal Capybara re-org
I am guessing this isn't an official public class for Capybara since
it was changed on a minor version (3.11). Don't want to yet assume we
want to drop pre-3.11 support so try loading both constants. Once we
are sure we want to drop anything less than 3.11 we can simplify to
just the new class name.